--- FocusStart-homeworks/TestTaskCFTTests/TestTaskCFTTests.swift
@@ -1,34 +1,0 @@
-//
-//  TestTaskCFTTests.swift
-//  TestTaskCFTTests
-//
-//  Created by Admin on 17.09.2020.
-//  Copyright © 2020 Admin. All rights reserved.
-//
-
-import XCTest
-@testable import TestTaskCFT
-
-class TestTaskCFTTests: XCTestCase {
-
-    override func setUpWithError() throws {
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-    }
-
-    override func tearDownWithError() throws {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-    }
-
-    func testExample() throws {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct results.
-    }
-
-    func testPerformanceExample() throws {
-        // This is an example of a performance test case.
-        self.measure {
-            // Put the code you want to measure the time of here.
-        }
-    }
-
-}
