+++ Task4/lekcia/Assets.xcassets/cute-dog.imageset/cute-dog.jpg
