+++ Task4/lekcia/Assets.xcassets/cute-rat.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "MW-IE203_corona_20200409232441_ZQ.jpg",
+      "filename" : "cute-rat.jpg",
       "idiom" : "universal",
       "scale" : "1x"
     },
