+++ Task4/lekcia/Controllers/DetailViewController/DetailViewController.swift
@@ -1,0 +1,51 @@
+//
+//  DetailViewController.swift
+//  lekcia
+//
+//  Created by Admin on 06.11.2020.
+//
+
+import UIKit
+
+class DetailViewController: UIViewController {
+	//static let identifier = String(describing: DetailViewController.self)
+	let roundedShadowImageView = RoundShadowImageView()
+	var text: String = ""
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+		view.backgroundColor = UIColor.red
+	  	configure()
+		self.navigationItem.title = "Detail"
+		self.navigationController?.navigationBar.prefersLargeTitles = true
+        // Do any additional setup after loading the view.
+    }
+
+
+	func configure() {
+		self.view.addSubview(roundedShadowImageView)
+		roundedShadowImageView.image = UIImage(named: "cute-dog")
+		NSLayoutConstraint.activate([roundedShadowImageView.trailingAnchor.constraint(
+										equalTo: view.safeAreaLayoutGuide.trailingAnchor,
+										constant: CGFloat(Constants.bottomAndTopSpace.rawValue)),
+									 roundedShadowImageView.trailingAnchor.constraint(
+										equalTo: view.safeAreaLayoutGuide.trailingAnchor,
+										constant: CGFloat(Constants.bottomAndTopSpace.rawValue)),
+									 roundedShadowImageView.trailingAnchor.constraint(
+										equalTo: view.safeAreaLayoutGuide.trailingAnchor,
+										constant: CGFloat(Constants.bottomAndTopSpace.rawValue)),
+									 roundedShadowImageView.trailingAnchor.constraint(
+										equalTo: view.safeAreaLayoutGuide.trailingAnchor,
+									 constant: CGFloat(Constants.bottomAndTopSpace.rawValue))])
+	}
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
