--- Task4/lekcia/Controllers/RoundShadowImageView.swift
@@ -1,8 +1,0 @@
-//
-//  RoundShadowImageView.swift
-//  lekcia
-//
-//  Created by Admin on 09.11.2020.
-//
-
-import Foundation
